video-ai-platform/worker/requirements.txt
# VIDEO AI WORKER - ALL 4 PHASES
# ================================

# Core ML frameworks
torch>=2.5.1
torchvision>=0.20.0
ultralytics>=8.3.0

# Computer Vision
opencv-python>=4.10.0
pillow>=11.1.0
numpy>=1.26.4
scipy>=1.14.0

# PHASE 1: CLIP for scene understanding (CRITICAL!)
git+https://github.com/openai/CLIP.git
ftfy
regex

# PHASE 2: Weighted Boxes Fusion ensemble
ensemble-boxes>=1.0.9

# PHASE 3: Panoptic Segmentation
transformers>=4.30.0

# Audio processing
openai-whisper>=1.0.0  # ✅ This is the right one!
librosa>=0.10.0           # ← ADD THIS! (for audio event detection)
soundfile>=0.12.1
torchaudio>=2.0.0
torch-audiomentations>=0.11.0

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0

# PHASE 4: Narrative Intelligence
anthropic>=0.39.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0
tqdm>=4.66.0

# Video processing
moviepy>=1.0.3
imageio>=2.35.0
imageio-ffmpeg>=0.5.0

# Data handling
pydantic>=2.9.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9

# Model tracking
tensorboard>=2.18.0